Do not add ~/bin to fish user PATHs in config.fish
First, the system is started as bash because of some reasons and then switched
to Fish so this way the ~/bin is not added to the system but only to the
terminal.

Second, I should use fish_add_path to avoid duplicates on each config.fish
execution. This solution is just plain wrong.